requirements.txt
tensorflow==2.16.1
tinkoff-investments==0.2.0b100
openpyxl==3.1.2
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
python-telegram-bot==21.2
tinkoff==0.1.1
pytest~=6.2.4
h11~=0.14.0
pip~=24.0
toml~=0.10.2
wheel~=0.43.0
cryptography~=42.0.5
py~=1.11.0
lxml~=5.2.2
tornado~=6.4
gast~=0.5.4
h5py~=3.11.0
ipython~=8.17.2
pytz~=2022.6
rich~=13.7.1
attrs~=23.2.0
pygments~=2.17.1
exceptiongroup~=1.1.3
sniffio~=1.3.1
click~=8.1.7
httpcore~=1.0.5
idna~=3.6
certifi~=2023.11.17
keras~=3.3.3
optree~=0.11.0
Pillow~=9.0.0
packaging~=24.0
scipy~=1.13.0
namex~=0.0.8
mdurl~=0.1.2
setuptools~=58.1.0
overrides~=7.7.0
cffi~=1.16.0
six~=1.16.0
matplotlib~=3.9.0
MarkupSafe~=2.1.5
python-dateutil~=2.8.2
cachetools~=5.3.3
deprecation~=2.1.0
urllib3~=1.26.18
anyio~=4.3.0
Markdown~=3.6
PyYAML~=6.0.1
defusedxml~=0.8.0rc2
APScheduler~=3.6.3
httpx~=0.27.0
termcolor~=2.4.0
flatbuffers~=24.3.25
wrapt~=1.16.0
astunparse~=1.6.3
tensorboard~=2.16.2
werkzeug~=3.0.1
zipp~=3.17.0
python-dotenv~=1.0.1
scikit-learn~=1.4.2